Added new reactive modules. Improved API and tests.
