+++ eventuate-tram-reactive-commands/src/main/java/io/eventuate/tram/reactive/commands/consumer/ReactiveCommandHandlerReplyBuilder.java
@@ -12,7 +12,7 @@ import reactor.core.publisher.Mono;
 
 public class ReactiveCommandHandlerReplyBuilder {
 
-  private static <T> Publisher<Message> with(T reply, CommandReplyOutcome outcome) {
+  private static <T> Mono<Message> with(T reply, CommandReplyOutcome outcome) {
     MessageBuilder messageBuilder = MessageBuilder
             .withPayload(JSonMapper.toJson(reply))
             .withHeader(ReplyMessageHeaders.REPLY_OUTCOME, outcome.name())
@@ -21,11 +21,11 @@ public class ReactiveCommandHandlerReplyBuilder {
     return Mono.defer(() -> Mono.just(messageBuilder.build()));
   }
 
-  public static Publisher<Message> withSuccess(Object reply) {
+  public static Mono<Message> withSuccess(Object reply) {
     return with(reply, CommandReplyOutcome.SUCCESS);
   }
 
-  public static Publisher<Message> withSuccess() {
+  public static Mono<Message> withSuccess() {
     return withSuccess(new Success());
   }
