+++ eventuate-tram-spring-reactive-consumer-kafka/src/main/java/io/eventuate/tram/spring/reactive/consumer/kafka/EventuateTramReactiveKafkaMessageConsumerConfiguration.java
@@ -5,13 +5,13 @@ import io.eventuate.messaging.kafka.spring.consumer.KafkaConsumerFactoryConfigur
 import io.eventuate.messaging.kafka.spring.consumer.MessageConsumerKafkaConfiguration;
 import io.eventuate.tram.consumer.common.reactive.ReactiveMessageConsumerImplementation;
 import io.eventuate.tram.reactive.consumer.kafka.EventuateTramReactiveKafkaMessageConsumer;
-import io.eventuate.tram.spring.reactive.consumer.common.ReactiveTramConsumerBaseCommonConfiguration;
+import io.eventuate.tram.spring.reactive.consumer.common.ReactiveTramConsumerCommonConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 
 @Configuration
-@Import({ReactiveTramConsumerBaseCommonConfiguration.class, MessageConsumerKafkaConfiguration.class, KafkaConsumerFactoryConfiguration.class})
+@Import({ReactiveTramConsumerCommonConfiguration.class, MessageConsumerKafkaConfiguration.class, KafkaConsumerFactoryConfiguration.class})
 public class EventuateTramReactiveKafkaMessageConsumerConfiguration {
 
   @Bean
