+++ eventuate-tram-spring-reactive-integration-tests/src/test/java/io/eventuate/tram/reactive/integrationtests/events/TestEventThatInitiatesException.java
@@ -1,15 +1,15 @@
-package io.eventuate.tram.reactive.integrationtests;
+package io.eventuate.tram.reactive.integrationtests.events;
 
 import io.eventuate.tram.events.common.DomainEvent;
 
-public class TestEvent implements DomainEvent {
+public class TestEventThatInitiatesException implements DomainEvent {
 
   private String payload;
 
-  public TestEvent() {
+  public TestEventThatInitiatesException() {
   }
 
-  public TestEvent(String payload) {
+  public TestEventThatInitiatesException(String payload) {
     this.payload = payload;
   }
