+++ eventuate-tram-spring-reactive-jdbc-kafka/build.gradle
@@ -1,0 +1,9 @@
+
+dependencies {
+    compile "io.eventuate.messaging.kafka:eventuate-messaging-kafka-spring-consumer:$eventuateMessagingKafkaVersion"
+
+    compile project(":eventuate-tram-spring-reactive-producer-jdbc")
+    compile project(":eventuate-tram-spring-reactive-consumer-common")
+    compile project(":eventuate-tram-spring-reactive-consumer-jdbc")
+    compile project(":eventuate-tram-spring-reactive-consumer-kafka")
+}
