+++ settings.gradle
@@ -35,6 +35,8 @@ include 'eventuate-tram-integration-test-common'
 include 'eventuate-tram-commands-db-broker-integration-test'
 include 'eventuate-tram-events'
 include 'eventuate-tram-reactive-events'
+include 'eventuate-tram-spring-reactive-events'
+include 'eventuate-tram-spring-reactive-jdbc-kafka'
 include 'eventuate-tram-spring-events-common'
 include 'eventuate-tram-spring-events-publisher'
 include 'eventuate-tram-spring-reactive-events-publisher'
